refactored the shopping list item component. #76
